+++ Source/Scenes/Main/MainBaseViewController.swift
@@ -146,7 +146,5 @@ class MainBaseViewController: UIViewController {
         let alertController = UIAlertController(title: "System Alert", message: msg, preferredStyle: .alert)
         alertController.addAction(UIAlertAction(title: "OK", style: .default))
         self.present(alertController, animated: true)
-
-        // TODO: handle ipad popoverController
     }
 }